包皮：版本管理

+++ app/src/main/java/com/habbyge/iwatch/patch/PatchManager.java
@@ -1,5 +1,6 @@
 package com.habbyge.iwatch.patch;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.util.Log;
@@ -31,41 +32,52 @@ public final class PatchManager {
     private static final String SUFFIX = ".ipatch"; // patch文件的后缀
     private static final String DIR = "ipatch";
     private static final String SP_NAME = "_iwatch_";
-    private static final String SP_VERSION = "ipatch_reversion";
+    private static final String SP_REVERSION = "ipatch_reversion";
 
-    private final Context mContext; // 这里必须是 Application 的 Context
+    private Context mContext; // 这里必须是 Application 的 Context
 
     /**
      * patch directory
      */
-    private final File mPatchDir;
+    private File mPatchDir;
     /**
      * patchs
      */
-    private final SortedSet<Patch> mPatchs;
+    private SortedSet<Patch> mPatchs; // TODO: 1/7/21 ing
     /**
      * classloaders
      */
-    private final Map<String, ClassLoader> mClassLoaderMap;
+    private Map<String, ClassLoader> mClassLoaderMap; // TODO: 1/7/21 ing
 
-    private final IWatch iWatch;
+    private IWatch iWatch;
 
-    /**
-     * @param context context
-     */
-    public PatchManager(Context context) {
-        mContext = context;
-        mPatchDir = new File(mContext.getFilesDir(), DIR);
-        iWatch = new IWatch(context);
-        // 线程安全的有序的集合，适用于高并发的场景
-        mPatchs = new ConcurrentSkipListSet<Patch>();
-        mClassLoaderMap = new ConcurrentHashMap<String, ClassLoader>();
+    @SuppressLint("StaticFieldLeak")
+    private static PatchManager mInstance;
+
+    public static PatchManager getInstance() {
+        if (mInstance == null) {
+            synchronized (PatchManager.class) {
+                if (mInstance == null) {
+                    mInstance = new PatchManager();
+                }
+            }
+        }
+        return mInstance;
+    }
+
+    private PatchManager() {
     }
 
     /**
+     * 初始化入口
+     *
+     * @param context 必须是全局的 Application 的 Context
      * @param appReversion App 打包粒度的 version
      */
-    public void init(String appReversion) {
+    public void init(Context context, String appReversion) {
+        mContext = context;
+
+        mPatchDir = new File(mContext.getFilesDir(), DIR);
         if (!mPatchDir.exists() && !mPatchDir.mkdirs()) {// make directory fail
             Log.e(TAG, "patch dir create error.");
             return;
@@ -74,20 +86,125 @@ public final class PatchManager {
             return;
         }
 
+        // 线程安全的有序的集合，适用于高并发的场景
+        mPatchs = new ConcurrentSkipListSet<Patch>();
+        mClassLoaderMap = new ConcurrentHashMap<String, ClassLoader>();
+
         SharedPreferences sp = mContext.getSharedPreferences(SP_NAME, Context.MODE_PRIVATE);
-        String reversion = sp.getString(SP_VERSION, null);
+        String reversion = sp.getString(SP_REVERSION, null);
         if (reversion == null || !reversion.equalsIgnoreCase(appReversion)) {
-            cleanPatch(); // 非对应 reversion 版本，则清理掉补丁，使用原始逻辑 TODO: ing......
-            sp.edit().putString(SP_VERSION, appReversion).apply();
+            cleanPatch(); // 非对应 reversion 版本，则清理掉补丁，使用原始逻辑
+            sp.edit().putString(SP_REVERSION, appReversion).apply();
+            Log.e(TAG, "PatchManager init failure appReversion=" + appReversion);
         } else {
             initPatchs();
         }
+
+        initIWatch(context);
     }
 
     private void initPatchs() {
         File[] files = mPatchDir.listFiles();
+        if (files == null) {
+            return;
+        }
         for (File file : files) {
             addPatch(file);
+        }
+        Log.i(TAG, "initPatchs success");
+    }
+
+    private void initIWatch(Context context) {
+        iWatch = new IWatch(context);
+        iWatch.init();
+    }
+
+    /**
+     * 实时打补丁的接口函数
+     * add patch at runtime
+     * @param path patch path
+     */
+    public void addPatch(String path) throws IOException {
+        File src = new File(path);
+        File dest = new File(mPatchDir, src.getName());
+        if (!src.exists()) {
+            throw new FileNotFoundException(path);
+        }
+        if (dest.exists()) {
+            Log.d(TAG, "patch [" + path + "] has be loaded.");
+            return;
+        }
+        FileUtil.copyFile(src, dest); // copy to patch's directory
+        Patch patch = addPatch(dest);
+        if (patch != null) {
+            loadPatch(patch);
+        }
+    }
+
+    /**
+     * remove all patchs
+     */
+    public void removeAllPatch() {
+        cleanPatch();
+        SharedPreferences sp = mContext.getSharedPreferences(SP_NAME, Context.MODE_PRIVATE);
+        sp.edit().clear().apply();
+    }
+
+    /**
+     * load patch,call when plugin be loaded. used for plugin architecture.</br>
+     * <p>
+     * need name and classloader of the plugin
+     *
+     * @param patchName   patch name
+     * @param classLoader classloader
+     */
+    public void loadPatch(String patchName, ClassLoader classLoader) {
+        mClassLoaderMap.put(patchName, classLoader);
+        Set<String> patchNames;
+        List<String> classes; // 该补丁文件(patchName)中所有的class
+        for (Patch patch : mPatchs) {
+            patchNames = patch.getPatchNames();
+            if (patchNames.contains(patchName)) {
+                classes = patch.getClasses(patchName);
+                iWatch.fix(patch.getFile(), classLoader, classes);
+            }
+        }
+    }
+
+    /**
+     * load patch, call when application start
+     */
+    public void loadPatch() {
+        mClassLoaderMap.put("*", mContext.getClassLoader()); // wildcard
+        Set<String> patchNames;
+        List<String> classes;
+        for (Patch patch : mPatchs) {
+            patchNames = patch.getPatchNames();
+            for (String patchName : patchNames) {
+                classes = patch.getClasses(patchName);
+                iWatch.fix(patch.getFile(), mContext.getClassLoader(), classes);
+            }
+        }
+    }
+
+    /**
+     * load specific patch
+     * @param patch patch
+     */
+    private void loadPatch(Patch patch) {
+        Set<String> patchNames = patch.getPatchNames();
+        ClassLoader classLoader;
+        List<String> classes;
+        for (String patchName : patchNames) {
+            if (mClassLoaderMap.containsKey("*")) {
+                classLoader = mContext.getClassLoader();
+            } else {
+                classLoader = mClassLoaderMap.get(patchName);
+            }
+            if (classLoader != null) {
+                classes = patch.getClasses(patchName);
+                iWatch.fix(patch.getFile(), classLoader, classes);
+            }
         }
     }
 
@@ -109,6 +226,9 @@ public final class PatchManager {
 
     private void cleanPatch() {
         File[] files = mPatchDir.listFiles();
+        if (files == null) {
+            return;
+        }
         for (File file : files) {
             iWatch.removeOptFile(file);
             if (!FileUtil.deleteFile(file)) {
@@ -117,97 +237,12 @@ public final class PatchManager {
         }
     }
 
-    /**
-     * 实时打补丁的接口函数
-     * add patch at runtime
-     * @param path patch path
-     */
-    @SuppressWarnings("unused")
-    public void addPatch(String path) throws IOException {
-        File src = new File(path);
-        File dest = new File(mPatchDir, src.getName());
-        if (!src.exists()) {
-            throw new FileNotFoundException(path);
-        }
-        if (dest.exists()) {
-            Log.d(TAG, "patch [" + path + "] has be loaded.");
-            return;
-        }
-        FileUtil.copyFile(src, dest);// copy to patch's directory
-        Patch patch = addPatch(dest);
-        if (patch != null) {
-            loadPatch(patch);
-        }
-    }
+    public void testFix(String className1, String funcName1, Class<?>[] paramTypes1,
+                        Class<?> returnType1, boolean isStatic1,
+                        String className2, String funcName2, Class<?>[] paramTypes2,
+                        Class<?> returnType2, boolean isStatic2) {
 
-    /**
-     * remove all patchs
-     */
-    @SuppressWarnings("unused")
-    public void removeAllPatch() {
-        cleanPatch();
-        SharedPreferences sp = mContext.getSharedPreferences(SP_NAME, Context.MODE_PRIVATE);
-        sp.edit().clear().apply();
-    }
-
-    /**
-     * load patch,call when plugin be loaded. used for plugin architecture.</br>
-     * <p>
-     * need name and classloader of the plugin
-     *
-     * @param patchName   patch name
-     * @param classLoader classloader
-     */
-    @SuppressWarnings("unused")
-    public void loadPatch(String patchName, ClassLoader classLoader) {
-        mClassLoaderMap.put(patchName, classLoader);
-        Set<String> patchNames;
-        List<String> classes;
-        for (Patch patch : mPatchs) {
-            patchNames = patch.getPatchNames();
-            if (patchNames.contains(patchName)) {
-                classes = patch.getClasses(patchName);
-                mAndFixManager.fix(patch.getFile(), classLoader, classes);
-            }
-        }
-    }
-
-    /**
-     * load patch, call when application start
-     */
-    @SuppressWarnings("unused")
-    public void loadPatch() {
-        mClassLoaderMap.put("*", mContext.getClassLoader());// wildcard
-        Set<String> patchNames;
-        List<String> classes;
-        for (Patch patch : mPatchs) {
-            patchNames = patch.getPatchNames();
-            for (String patchName : patchNames) {
-                classes = patch.getClasses(patchName);
-                mAndFixManager.fix(patch.getFile(), mContext.getClassLoader(), classes);
-            }
-        }
-    }
-
-    /**
-     * load specific patch
-     * @param patch patch
-     */
-    private void loadPatch(Patch patch) {
-        Set<String> patchNames = patch.getPatchNames();
-        ClassLoader classLoader;
-        List<String> classes;
-        for (String patchName : patchNames) {
-            if (mClassLoaderMap.containsKey("*")) {
-                classLoader = mContext.getClassLoader();
-            } else {
-                classLoader = mClassLoaderMap.get(patchName);
-            }
-            if (classLoader != null) {
-                classes = patch.getClasses(patchName);
-                mAndFixManager.fix(patch.getFile(), classLoader, classes);
-            }
-        }
+        iWatch.hook(className1, funcName1, paramTypes1, returnType1, isStatic1,
+                    className2, funcName2, paramTypes2, returnType2, isStatic2);
     }
 }
-
